Atualizando Feed para Global State
